// src/Application/ApiCalls/generated/models/DrapeauxOutput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AXECIEL.GMAO.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DrapeauxOutput
 */
export interface DrapeauxOutput {
    /**
     * 
     * @type {string}
     * @memberof DrapeauxOutput
     */
    noIntervention?: string | null;
    /**
     * 
     * @type {string}
     * @memberof DrapeauxOutput
     */
    codeImmeuble?: string | null;
    /**
     * 
     * @type {string}
     * @memberof DrapeauxOutput
     */
    adressImmeuble?: string | null;
    /**
     * 
     * @type {string}
     * @memberof DrapeauxOutput
     */
    designation?: string | null;
    /**
     * 
     * @type {string}
     * @memberof DrapeauxOutput
     */
    dateRealise?: string | null;
}

/**
 * Check if a given object implements the DrapeauxOutput interface.
 */
export function instanceOfDrapeauxOutput(value: object): value is DrapeauxOutput {
    return true;
}

export function DrapeauxOutputFromJSON(json: any): DrapeauxOutput {
    return DrapeauxOutputFromJSONTyped(json, false);
}

export function DrapeauxOutputFromJSONTyped(json: any, ignoreDiscriminator: boolean): DrapeauxOutput {
    if (json == null) {
        return json;
    }
    return {
        
        'noIntervention': json['noIntervention'] == null ? undefined : json['noIntervention'],
        'codeImmeuble': json['codeImmeuble'] == null ? undefined : json['codeImmeuble'],
        'adressImmeuble': json['adressImmeuble'] == null ? undefined : json['adressImmeuble'],
        'designation': json['designation'] == null ? undefined : json['designation'],
        'dateRealise': json['dateRealise'] == null ? undefined : json['dateRealise'],
    };
}

export function DrapeauxOutputToJSON(json: any): DrapeauxOutput {
    return DrapeauxOutputToJSONTyped(json, false);
}

export function DrapeauxOutputToJSONTyped(value?: DrapeauxOutput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'noIntervention': value['noIntervention'],
        'codeImmeuble': value['codeImmeuble'],
        'adressImmeuble': value['adressImmeuble'],
        'designation': value['designation'],
        'dateRealise': value['dateRealise'],
    };
}