// src/Application/ApiCalls/generated/models/DeleteDrapeauxInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AXECIEL.GMAO.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DeleteDrapeauxInput
 */
export interface DeleteDrapeauxInput {
    /**
     * 
     * @type {Array<string>}
     * @memberof DeleteDrapeauxInput
     */
    noIntervention?: Array<string> | null;
}

/**
 * Check if a given object implements the DeleteDrapeauxInput interface.
 */
export function instanceOfDeleteDrapeauxInput(value: object): value is DeleteDrapeauxInput {
    return true;
}

export function DeleteDrapeauxInputFromJSON(json: any): DeleteDrapeauxInput {
    return DeleteDrapeauxInputFromJSONTyped(json, false);
}

export function DeleteDrapeauxInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): DeleteDrapeauxInput {
    if (json == null) {
        return json;
    }
    return {
        
        'noIntervention': json['noIntervention'] == null ? undefined : json['noIntervention'],
    };
}

export function DeleteDrapeauxInputToJSON(json: any): DeleteDrapeauxInput {
    return DeleteDrapeauxInputToJSONTyped(json, false);
}

export function DeleteDrapeauxInputToJSONTyped(value?: DeleteDrapeauxInput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'noIntervention': value['noIntervention'],
    };
}